requirements.txt
PyMuPDF>=1.23.0
google-generativeai>=0.3.0

